Add more logging (#159)

+++ src/main/java/uk/gov/hmcts/reform/draftstore/service/DraftService.java
@@ -1,5 +1,7 @@
 package uk.gov.hmcts.reform.draftstore.service;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.draftstore.data.DraftStoreDao;
 import uk.gov.hmcts.reform.draftstore.domain.CreateDraft;
@@ -12,6 +14,7 @@ import uk.gov.hmcts.reform.draftstore.exception.NoDraftFoundException;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.toList;
 import static uk.gov.hmcts.reform.draftstore.service.mappers.FromDbModelMapper.fromDb;
@@ -19,6 +22,8 @@ import static uk.gov.hmcts.reform.draftstore.service.mappers.ToDbModelMapper.toD
 
 @Service
 public class DraftService {
+
+    private static final Logger log = LoggerFactory.getLogger(DraftService.class);
 
     private final DraftStoreDao draftRepo;
 
@@ -47,11 +52,15 @@ public class DraftService {
     }
 
     public int create(CreateDraft newDraft, UserAndService userAndService) {
-        return draftRepo.insert(
+        int id = draftRepo.insert(
             userAndService.userId,
             userAndService.service,
             toDb(newDraft, userAndService.secrets)
         );
+
+        log.info("Created draft. ID: {}, service: {}", id, userAndService.service);
+
+        return id;
     }
 
     public void update(String id, UpdateDraft updatedDraft, UserAndService userAndService) {
@@ -61,22 +70,27 @@ public class DraftService {
         if (draft.isPresent()) {
             assertCanEdit(draft.get(), userAndService);
             draftRepo.update(toInternalId(id), toDb(updatedDraft, userAndService.secrets));
+            log.info("Updated draft. ID: {}", id);
         } else {
             throw new NoDraftFoundException();
         }
     }
 
     public void delete(String id, UserAndService userAndService) {
-        draftRepo
-            .read(toInternalId(id))
-            .ifPresent(d -> {
-                assertCanEdit(d, userAndService);
-                draftRepo.delete(toInternalId(id));
-            });
+        Optional<uk.gov.hmcts.reform.draftstore.data.model.Draft> draft = draftRepo.read(toInternalId(id));
+
+        if (draft.isPresent()) {
+            assertCanEdit(draft.get(), userAndService);
+            draftRepo.delete(toInternalId(id));
+            log.info("Deleted draft. ID: {}", id);
+        } else {
+            log.warn("Tried to delete draft {}, but it was not found", id);
+        }
     }
 
     public void deleteAll(UserAndService userAndService) {
         draftRepo.deleteAll(userAndService.userId, userAndService.service);
+        log.info("Deleted all drafts for user {} in service {}", userAndService.userId, userAndService.service);
     }
 
     private void assertCanEdit(uk.gov.hmcts.reform.draftstore.data.model.Draft draft, UserAndService userAndService) {
